step16 DeleteCoursesDetails

+++ TD2-SOAA-master/TD2-SOAA-master/src/main/java/org/example/course_details/GetCourseDetailsRequest.java
@@ -2,7 +2,7 @@
 // Ce fichier a été généré par l'implémentation de référence JavaTM Architecture for XML Binding (JAXB), v2.2.7 
 // Voir <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Toute modification apportée à ce fichier sera perdue lors de la recompilation du schéma source. 
-// Généré le : 2020.11.12 à 12:04:17 PM UTC 
+// Généré le : 2020.11.17 à 12:11:45 PM UTC 
 //
 
 
@@ -43,7 +43,7 @@ import javax.xml.bind.annotation.XmlType;
 public class GetCourseDetailsRequest {
 
     @XmlElement(required = true)
-    protected int id;
+    protected BigInteger id;
 
     /**
      * Obtient la valeur de la propriété id.
@@ -53,7 +53,7 @@ public class GetCourseDetailsRequest {
      *     {@link BigInteger }
      *     
      */
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
 
@@ -65,7 +65,7 @@ public class GetCourseDetailsRequest {
      *     {@link BigInteger }
      *     
      */
-    public void setId(int value) {
+    public void setId(BigInteger value) {
         this.id = value;
     }
